Fix EEPROM write operation to store password correctly
The code change modifies the `handleSubmit` function in the `Ethernet_to_serial.ino` file. Specifically, it updates the call to the `writeString` function. Previously, the function was writing the `storedUsername` to the `EEP_USERNAME` EEPROM address. The change updates this to write the `storedPassword` to the `EEP_PASSWORD` EEPROM address instead. This ensures that the password is correctly stored in the EEPROM.